[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2024 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -6,13 +6,13 @@ using namespace std;
 
 int main() {
 
-	int l , r = 31, result =0;
+	long long l , r = 31, result =0;
 	string str;
 
 	cin >> l >> str;
 
 	for (int i = 0; i < str.length(); i++) {
-		int tmp = str[i] - 'a' + 1;
+		long long int tmp = str[i] - 'a' + 1;
 		for (int j = 0; j < i; j++) {
 			tmp *= r;
 			tmp %= 1234567891;
